+++ api/node/client.js
@@ -12,7 +12,6 @@ const DIController = require('../../shared/DIController')
  *  id: String,
  *  role: Number,
  *  heartbeat: Number,
- *  isPersistent: Boolean,
  *  isEditingLayout: Boolean
  * }} Connection
  */