+++ app/App.jsx
@@ -1,8 +1,10 @@
 import React from 'react'
 import { Start } from './views/Start'
 import { Workspace } from './views/Workspace'
+import { WorkspaceWidget } from './views/WorkspaceWidget'
 
 import { Router } from './components/Router'
+import { Transparency } from './components/Transparency'
 
 import { LocalContext } from './localContext'
 import { SharedContext } from './sharedContext'
@@ -12,28 +14,15 @@ import { useWebsocket } from './hooks/useWebsocket'
 
 import * as shortcuts from './utils/shortcuts'
 import * as browser from './utils/browser'
+import * as auth from './auth'
 import * as api from './api'
-
-/**
- * Define the interval of heartbeats
- * sent to the server to indicate that
- * the socket is alive
- *
- * This value MUST be smaller than the
- * ttl of a socket defined in the server
- *
- * Defaults to 5 seconds
- *
- * @type { Number }
- */
-const HEARTBEAT_INTERVAL_MS = 5000
 
 /**
   * The protocol (wss or ws)
   * that sockets should use
   * based on the current http
   * protocol
-  * @type { String }
+  * @type { string }
   */
 const socketProtocol = (function () {
   if (window.location.protocol === 'https:') {
@@ -79,11 +68,15 @@ root html tag for platform-specific styling e.t.c.
   window.document.documentElement.dataset.platform = browser.platform()
 })()
 
+const websocketQuery = {
+  workspace
+}
+
 export default function App () {
   const [local, setLocal] = React.useState({})
   const [shared, setShared] = React.useState({})
 
-  const [data, send, readyState] = useWebsocket(`${socketHost}/api/v1/ws?workspace=${workspace}`, true)
+  const [data, send, readyState] = useWebsocket(workspace && `${socketHost}/api/v1/ws`, true, websocketQuery)
 
   /**
     * Setup a reference to hold
@@ -99,11 +92,6 @@ export default function App () {
   }, [local])
 
   React.useEffect(() => {
-    if (readyState !== 1) return
-    send({ type: 'id', data: local.id })
-  }, [readyState])
-
-  React.useEffect(() => {
     /**
      * Setup the Bridge api
      * and attach listeners
@@ -114,6 +102,12 @@ export default function App () {
       bridge.transport.send = msg => {
         send(msg)
       }
+
+      if (!bridge.client.getIdentity()) {
+        const id = await bridge.client.registerClient()
+        applyLocal({ id })
+      }
+      
       bridge.transport.replayQueue()
 
       bridge.events.on('state.change', state => {
@@ -121,7 +115,7 @@ export default function App () {
       })
 
       window.onbeforeunload = () => {
-        send({ type: 'disconnect' })
+        bridge.client.removeClient()
       }
 
       const initialState = await bridge.state.get()
@@ -131,25 +125,15 @@ export default function App () {
     setup()
   }, [readyState])
 
-  /*
-  Setup an interval to send a heartbeat
-  at a regular interval to the server
-  to indicate that the socket is alive
-  */
   React.useEffect(() => {
-    async function sendHeartbeat () {
+    async function setup () {
+      const token = await auth.getToken()
       const bridge = await api.load()
-      bridge.client.heartbeat()
+      bridge.commands.setHeader('authentication', token)
     }
-
-    const ival = setInterval(
-      () => sendHeartbeat(),
-      HEARTBEAT_INTERVAL_MS
-    )
-    sendHeartbeat()
-
-    return () => clearInterval(ival)
-  }, [local])
+    if (readyState !== 1) return
+    setup()
+  }, [readyState])
 
   /**
    * Apply data to the shared state,
@@ -187,37 +171,11 @@ export default function App () {
   */
   React.useEffect(() => {
     ;(async function () {
-      if (!data) return
-      const json = JSON.parse(data)
-      switch (json?.type) {
-        /*
-        Keep track of this connection's
-        unique identifier and setup the
-        client's initial state
-        */
-        case 'id':
-          applyLocal({ id: json?.data })
-          applyShared({
-            _connections: {
-              [json?.data]: {
-                isPersistent: browser.isElectron()
-              }
-            }
-          })
-          ;(await api.load()).client.setIdentity(json?.data)
-          break
-
-        /*
-        Forward the message to
-        the api for processing
-        */
-        default:
-          ;(async function () {
-            const bridge = await api.load()
-            bridge.transport.receive(json)
-          })()
-          break
+      if (!data) {
+        return
       }
+      const bridge = await api.load()
+      bridge.transport.receive(data)
     })()
   }, [data])
 
@@ -245,7 +203,12 @@ export default function App () {
     <SocketContext.Provider value={[send, data]}>
       <LocalContext.Provider value={[local, applyLocal]}>
         <SharedContext.Provider value={[shared, applyShared]}>
+          <Transparency />
           <Router routes={[
+            {
+              path: /^\/workspaces\/.+\/widgets\/.+$/,
+              render: () => <WorkspaceWidget />
+            },
             {
               path: /^\/workspaces\/.+$/,
               render: () => <Workspace />