+++ app/assets/icons/index.js
@@ -5,7 +5,10 @@
 import add from './add.svg'
 import edit from './edit.svg'
 import close from './close.svg'
+import float from './float.svg'
 import person from './person.svg'
+import reload from './reload.svg'
+import search from './search.svg'
 import widget from './widget.svg'
 import rundown from './rundown.svg'
 import warning from './warning.svg'
@@ -23,7 +26,10 @@ export default {
   add,
   edit,
   close,
+  float,
   person,
+  reload,
+  search,
   widget,
   rundown,
   warning,